week2/day5/school_management/requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0